Avg calculation

+++ backend/src/main/java/com/book/backend/Models/Book.java
@@ -33,7 +33,7 @@ public class Book {
     private String edition ;
     private String isbn;
     private Integer sold;
-    private Integer avg;
+    private Double avg;
     private List<String> reviewIds;
     private List<String> imageIds;
     private String language;
